requirements.txt
asgiref==3.5.0
astropy==5.0.1
astroquery==0.4.5
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-crispy-forms==1.14.0
django-filter==21.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
html5lib==1.1
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.1.1
keyring==23.5.0
Markdown==3.3.6
MarkupSafe==2.1.1
mongoengine==0.24.1
numpy==1.22.2
packaging==21.3
pyerfa==2.0.0.1
pymongo==3.12.1
pyparsing==3.0.7
pytz==2021.3
pyvo==1.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
wincertstore==0.2
zipp==3.7.0
